Move the isKeyRegexLike implementation in DataOutputs to avoid polluting UtilityRoutines
