+++ src/EnergyPlus/DataOutputs.cc
@@ -135,8 +135,35 @@ Array1D_string const MonthlyNamedReports(NumMonthlyReports,
                                           "MECHANICALVENTILATIONLOADSMONTHLY",
                                           "HEATEMISSIONSREPORTMONTHLY"});
 
+bool isKeyRegexLike(std::string_view key)
+{
+    if (key == "*") {
+        return false;
+    }
+
+    return key.find_first_of("*+?()|[]\\.") != std::string_view::npos;
+}
+
+bool isKeyRegexLikeOri(std::string_view key)
+{
+    // DataOutputs does it like this
+    if (key == "*") {
+        return false;
+    }
+
+    bool is_simple_string = true;
+    for (auto const &c : key) {
+        if (c == ' ' || c == '_' || (std::isalnum(c) != 0)) {
+            continue;
+        }
+        is_simple_string = false;
+        break;
+    }
+    return !is_simple_string;
+}
+
 OutputReportingVariables::OutputReportingVariables(EnergyPlusData &state, std::string const &KeyValue, std::string const &VariableName)
-    : key(KeyValue), variableName(VariableName), is_simple_string(!UtilityRoutines::isKeyRegexLikeOri(KeyValue))
+    : key(KeyValue), variableName(VariableName), is_simple_string(!DataOutputs::isKeyRegexLikeOri(KeyValue))
 {
 
     if (is_simple_string) {
