+++ src/EnergyPlus/UtilityRoutines.hh
@@ -626,10 +626,6 @@ namespace UtilityRoutines {
                                                    DataGlobalConstants::ResourceType &FuelTypeNum,
                                                    bool &FuelTypeErrorsFound);
 
-    // TODO: keep only one of these two
-    bool isKeyRegexLike(std::string_view key);
-    bool isKeyRegexLikeOri(std::string_view key); // DataOutputs implementation
-
 } // namespace UtilityRoutines
 
 constexpr int getEnumerationValue(const gsl::span<const std::string_view> sList, const std::string_view s)
